refactor: Move test_{teamocil,tmuxinator}.py -> workspace/test_{\1}.py
